+++ packages/stream-model-handler/src/__tests__/tile-document.test.ts
@@ -5,7 +5,7 @@ import * as dagCBOR from '@ipld/dag-cbor'
 import type { DID } from 'dids'
 import { wrapDocument } from '@ceramicnetwork/3id-did-resolver'
 import * as KeyDidResolver from 'key-did-resolver'
-import { ModelHandler } from '../tile-document-handler.js'
+import { ModelHandler } from '../model-handler.js'
 import * as uint8arrays from 'uint8arrays'
 import * as sha256 from '@stablelib/sha256'
 import cloneDeep from 'lodash.clonedeep'
