+++ packages/stream-model-handler/src/model-handler.ts
