+++ packages/stream-model/src/model.ts
